requirements.txt
Django==1.6
South==0.8.3
beautifulsoup4==4.3.2
django-pagination==1.0.7
django-registration==1.0
requests==2.0.1
wsgiref==0.1.2
